+++ src/struct-buffer.ts
@@ -1,20 +1,20 @@
 import { StructType } from "./class-type";
 import { AnyObject } from "./interfaces";
 import {
-  arrayNextProxy,
-  arrayProxy,
+  arrayProxyNext,
   createDataView,
   makeDataView,
   unflattenDeep,
   zeroMemory,
 } from "./utils";
 
+export type Type_t = StructType<any, any> | StructBuffer;
+export type StructBuffer_t = { [k: string]: Type_t };
+
 /**
  * Get the size after byte alignment
  */
-export function sizeof(
-  type: StructType<any, any> | StructBuffer<IStructBuffer>
-): number {
+export function sizeof(type: Type_t): number {
   if (type instanceof StructBuffer) {
     let padidng = 0;
     const maxSize = type.maxSize;
@@ -25,7 +25,7 @@ export function sizeof(
   return type.isList ? type.size * type.count : type.size;
 }
 
-function byteLength(sb: StructBuffer<IStructBuffer>, count?: number) {
+function byteLength(sb: StructBuffer, count?: number) {
   const typeByteLength: number = Object.values(sb.struct).reduce(
     (acc: number, type) => {
       if (type instanceof StructBuffer) acc += type.byteLength;
@@ -37,35 +37,50 @@ function byteLength(sb: StructBuffer<IStructBuffer>, count?: number) {
   return typeByteLength * (count ?? sb.count);
 }
 
-export interface IStructBuffer {
-  [k: string]: StructType<any, any> | StructBuffer<IStructBuffer>;
-}
-
-class StructBufferNext<T extends IStructBuffer> {
-  deeps: number[] = [];
-  constructor(
-    i: number,
-    public readonly structName: string,
-    public readonly struct: T
-  ) {
-    this.deeps.push(i);
-    return arrayNextProxy(this);
+class StructBufferNext {
+  constructor() {
+    return arrayProxyNext(this, StructBufferNext);
   }
 }
+
+type StructBufferConfig = {
+  textDecode?: TextDecoder;
+  textEncoder?: TextEncoder;
+
+  /**
+   * Setting littleEndian here will cause the littleEndian parameters of `encode` and `decode` to become invalid
+   * 
+   * https://github.com/januwA/struct-buffer/issues/2
+   */
+  littleEndian?: boolean;
+};
+
+const KStructBufferConfig = {
+  textDecode: new TextDecoder(),
+  textEncoder: new TextEncoder(),
+  littleEndian: undefined,
+};
 
 export class StructBuffer<
-  T extends IStructBuffer,
   D = {
-    [k in keyof T]: any;
+    [k in keyof StructBuffer_t]: any;
   },
-  E = {
-    [k in keyof T]?: any;
-  }
-> extends Array<StructBuffer<T, D[], E[]>> {
+  E = Partial<D>
+> extends Array<StructBuffer<D[], E[]>> {
   deeps: number[] = [];
-  textDecode = new TextDecoder();
-  textEncoder = new TextEncoder();
-  structKV: [string, StructType<any, any> | StructBuffer<IStructBuffer>][];
+  config: StructBufferConfig = Object.assign({}, KStructBufferConfig);
+  structKV: [string, Type_t][];
+
+  constructor(
+    public structName: string,
+    public struct: StructBuffer_t,
+    config?: StructBufferConfig
+  ) {
+    super();
+    Object.assign(this.config, config);
+    this.structKV = Object.entries(struct);
+    return arrayProxyNext(this, StructBufferNext);
+  }
 
   /**
    *
@@ -88,19 +103,6 @@ export class StructBuffer<
     return this.deeps.reduce((acc, it) => (acc *= it), 1);
   }
 
-  constructor(public readonly structName: string, public readonly struct: T) {
-    super();
-    this.structKV = Object.entries(struct);
-    return arrayProxy(this, (o, i) => {
-      const newProxy: any = new StructBufferNext(i, o.structName, o.struct);
-      newProxy.textDecode = o.textDecode;
-      newProxy.textEncoder = o.textEncoder;
-      newProxy.structKV = o.structKV;
-      Object.setPrototypeOf(newProxy, StructBuffer.prototype);
-      return newProxy;
-    });
-  }
-
   get byteLength(): number {
     return byteLength(this as any);
   }
@@ -118,20 +120,21 @@ export class StructBuffer<
     littleEndian: boolean = false,
     offset: number = 0
   ): D {
+    littleEndian = this.config.littleEndian ?? littleEndian;
     view = makeDataView(view);
     const result: AnyObject[] = [];
     let i = this.count;
     while (i--) {
       const data = this.structKV.reduce<AnyObject>((acc, [key, type]) => {
         if (type instanceof StructBuffer) {
-          acc[key] = type.decode(view, littleEndian, offset);
+          acc[key] = type.decode(view, type.config.littleEndian ?? littleEndian, offset);
           offset += type.byteLength;
         } else {
           acc[key] = (type as any).decode(
             view,
             littleEndian,
             offset,
-            this.textDecode
+            this.config.textDecode
           );
           offset += sizeof(type);
         }
@@ -149,8 +152,8 @@ export class StructBuffer<
     offset: number = 0,
     view?: DataView
   ): DataView {
+    littleEndian = this.config.littleEndian ?? littleEndian;
     const v = createDataView(this.byteLength, view);
-
     if (this.isList && Array.isArray(obj)) (obj as any) = obj.flat();
 
     for (let i = 0; i < this.count; i++) {
@@ -161,22 +164,22 @@ export class StructBuffer<
         offset += itemSize;
         continue;
       }
-      this.structKV.reduce<DataView>((acc: DataView, [key, type]) => {
+      this.structKV.reduce<DataView>((view: DataView, [key, type]) => {
         const value = it[key];
         if (type instanceof StructBuffer) {
-          type.encode(value, littleEndian, offset, acc);
+          type.encode(value, type.config.littleEndian ?? littleEndian, offset, view);
           offset += type.byteLength;
         } else {
           (type as any).encode(
             value,
             littleEndian,
             offset,
-            acc,
-            this.textEncoder
+            view,
+            this.config.textEncoder
           );
           offset += sizeof(type);
         }
-        return acc;
+        return view;
       }, v);
     }
     return v;